# Conflicts:
#	src/app/features/attendance/showAttendance/presentation/dialogAttendance-details/dialogAttendance-details.html
#	src/app/features/definitions/Penalties/presentation/Penalties.component.html
#	src/app/features/definitions/Penalties/presentation/Penalties.component.ts
#	src/app/features/definitions/bank-branches/bank-branches.facade.ts
#	src/app/features/definitions/bank-branches/bank-branches.interface.ts
#	src/app/features/definitions/bank-branches/bank-branches.services.ts
#	src/app/features/definitions/bank-branches/presentation/bank-branches.component.html
#	src/app/features/definitions/bank/banks.facade.ts
#	src/app/features/definitions/bank/banks.services.ts
#	src/app/features/definitions/bank/presentation/banks.component.html
#	src/app/features/definitions/bank/presentation/banks.component.ts
#	src/app/features/definitions/classification-bankBranches/classification-bankBranches.facade.ts
#	src/app/features/definitions/classification-bankBranches/classification-bankBranches.interface.ts
#	src/app/features/definitions/classification-bankBranches/classification-bankBranches.services.ts
#	src/app/features/definitions/classification-bankBranches/presentation/classification-bankBranches.component.html
#	src/app/features/definitions/classification-bankBranches/presentation/classification-bankBranches.component.ts
#	src/app/features/definitions/courts/presentation/courts.component.html
#	src/app/features/definitions/courts/presentation/courts.component.ts
#	src/app/features/definitions/document-types/document-types.facade.ts
#	src/app/features/definitions/document-types/document-types.services.ts
#	src/app/features/definitions/document-types/presentation/document-types.component.html
#	src/app/features/definitions/evaluations-types/presentation/evaluations-types.component.ts
#	src/app/features/definitions/nationalities/presentation/nationalities.component.html
#	src/app/features/definitions/nationalities/presentation/nationalities.component.ts
#	src/app/features/definitions/rewards-types/presentation/rewards-types.component.html
#	src/app/features/definitions/rewards-types/presentation/rewards-types.component.ts
#	src/app/features/definitions/rewards-types/rewards-types.facade.ts
#	src/app/features/definitions/rewards-types/rewards-types.services.ts
#	src/app/features/definitions/scientific-qualifications/presentation/scientific-qualifications.component.html
#	src/app/features/definitions/scientific-qualifications/presentation/scientific-qualifications.component.ts
#	src/app/features/definitions/vacations-types/presentation/vacations-types.component.html
#	src/app/features/definitions/vacations-types/vacations-types.facade.ts
#	src/app/features/definitions/vacations-types/vacations-types.interface.ts
#	src/app/features/definitions/vacations-types/vacations-types.services.ts

+++ src/app/features/administrativeAffairs/employeeEvaluation/presentation/employee-evaluation.component.ts
@@ -1,11 +1,8 @@
-import {ChangeDetectorRef, Component, OnInit} from '@angular/core';
-import {FormBuilder, Validators} from "@angular/forms";
-import {optionsBooleanGeneral, optionsEvaluation, optionsJobClassification} from "../../../../core/core.interface";
-import {OrganizationalUnitFacade} from "../../organizational-unit/organizational-unit.facade";
-import {EmployeeEvaluationFacade} from "../employee-evaluation.facade";
-import {EmployeeFacade} from "../../employee/employee.facade";
-
-declare var $: any;
+import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { FormBuilder, Validators } from '@angular/forms';
+import { optionsEvaluation } from '../../../../core/core.interface';
+import { EmployeeEvaluationFacade } from '../employee-evaluation.facade';
+import { EmployeeFacade } from '../../employee/employee.facade';
 
 @Component({
   selector: 'app-rewards-types',
@@ -20,35 +17,50 @@ export class EmployeeEvaluationComponent implements OnInit {
     year: [0, Validators.required],
     evaluationId: [0, Validators.required],
     evaluationName: [''],
-    employeeName: [''],
+    employeeName: ['']
   });
   registerFormSearch = this.fb.group({
-    employeeId: ['', Validators.required],
+    employeeId: ['', Validators.required]
   });
-  constructor(  private fb: FormBuilder,
-                protected employeeFacade: EmployeeFacade,
-                protected employeeEvaluationFacade: EmployeeEvaluationFacade,
-                private cdr: ChangeDetectorRef) {
-    this.onSubmit();
-    this.employeeFacade.GetEmployee();
+
+  loadEmployees = (page: number, pageSize: number, searchQuery?: string): void => {
+    this.employeeFacade.GetEmployee(page, pageSize);
+  };
+
+  onEmployeeSelect(employee: any) {
+    this.registerForm.controls.employeeId.setValue(employee.id);
+  }
+
+  onEmployeeSelectRegisterFormSearch(employee: any) {
+    this.registerFormSearch.controls.employeeId.setValue(employee.id);
+  }
+
+  constructor(
+    private fb: FormBuilder,
+    protected employeeFacade: EmployeeFacade,
+    protected employeeEvaluationFacade: EmployeeEvaluationFacade,
+    private cdr: ChangeDetectorRef
+  ) {
+    this.employeeEvaluationFacade.GetEmployeeEvaluation(null);
   }
   ngOnInit() {
     this.edit = false;
+    this.loadEmployees(1, 10);
   }
-  onSubmit(): void {
-    this.employeeEvaluationFacade.GetEmployeeEvaluation(null);
-  }
+
   onSearch(): void {
     this.registerForm.controls.id.setValue('');
-if (this.registerFormSearch.valid){
+    if (this.registerFormSearch.valid) {
       this.employeeEvaluationFacade.GetEmployeeEvaluation(this.registerFormSearch.value.employeeId);
-  }
+    }
   }
 
   onDelete(Id: string): void {
-    this.edit = false;
-    this.employeeEvaluationFacade.deleteEmployeeEvaluation(Id);
-    this.registerForm.reset();
+    if (confirm('هل أنت متأكد من عملية المسح؟')) {
+      this.edit = false;
+      this.employeeEvaluationFacade.deleteEmployeeEvaluation(Id);
+      this.registerForm.reset();
+    }
   }
   onReset(): void {
     this.edit = false;
@@ -58,26 +70,29 @@ if (this.registerFormSearch.valid){
     this.registerFormSearch.setErrors(null);
   }
   onAdd(): void {
-    this.registerForm.value.evaluationName = this.optionsEvaluation.find(option => option.value == this.registerForm.value.evaluationId)?.label;
-    const optionEmployee = this.employeeFacade.employeeSubject$.getValue().find(x => x.id == this.registerForm.value.employeeId);
-    this.registerForm.value.employeeName =  this.registerForm.value.employeeId != '' && this.registerForm.value.employeeId != null ?   optionEmployee.name: '';
+    this.registerForm.value.evaluationName = this.optionsEvaluation.find(
+      (option) => option.value == this.registerForm.value.evaluationId
+    )?.label;
+    const optionEmployee = this.employeeFacade.employeeSubject$.getValue().items.find((x) => x.id == this.registerForm.value.employeeId);
+    this.registerForm.value.employeeName =
+      this.registerForm.value.employeeId != '' && this.registerForm.value.employeeId != null ? optionEmployee.name : '';
     if (this.registerForm.valid) {
-      if(this.edit) {
+      if (this.edit) {
         this.employeeEvaluationFacade.UpdateEmployeeEvaluation(this.registerForm?.value);
         this.onReset();
-      }else{
+      } else {
         this.employeeEvaluationFacade.AddEmployeeEvaluation(this.registerForm?.value);
         this.onReset();
-
       }
     }
   }
   onEdit(jobTitle: any): void {
     this.registerForm.patchValue(jobTitle);
-    this.registerForm.value.evaluationName = this.optionsEvaluation.find(option => option.value == this.registerForm.value.evaluationId)?.label;
+    this.registerForm.value.evaluationName = this.optionsEvaluation.find(
+      (option) => option.value == this.registerForm.value.evaluationId
+    )?.label;
     this.edit = true;
-
   }
 
-  protected readonly optionsEvaluation  = optionsEvaluation;
+  protected readonly optionsEvaluation = optionsEvaluation;
 }
