+++ src/app/features/definitions/Penalties/presentation/Penalties.component.ts
@@ -1,10 +1,9 @@
-import { Component, OnInit} from '@angular/core';
-import {FormBuilder, Validators} from "@angular/forms";
+import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { FormBuilder, Validators } from '@angular/forms';
 import { PenaltiesFacade } from '../Penalties.facade';
 import { optionsPenaltyType } from '../Penalties.interface';
 import { MessageType } from '../../../../shared/shared.interfaces';
 import { SharedFacade } from '../../../../shared/shared.facade';
-declare var $: any;
 
 @Component({
   selector: 'app-rewards-types',
@@ -17,12 +16,12 @@ export class PenaltiesComponent implements OnInit {
     id: [''],
     penaltyTypeId: [null, Validators.required],
     discount: [null, Validators.required],
-    penaltyTypeName: [''],
+    penaltyTypeName: ['']
   });
   constructor(
-      private fb: FormBuilder,
-      protected penaltiesFacade: PenaltiesFacade,
-      private sharedFacade: SharedFacade
+    private fb: FormBuilder,
+    protected penaltiesFacade: PenaltiesFacade,
+    private sharedFacade: SharedFacade
   ) {
     this.onSubmit();
   }
@@ -35,9 +34,11 @@ export class PenaltiesComponent implements OnInit {
     this.penaltiesFacade.GetPenalties();
   }
   onDelete(Id: string): void {
-    this.edit = false;
-    this.penaltiesFacade.deletePenalties(Id);
-    this.registerForm.reset();
+    if (confirm('هل أنت متأكد من عملية المسح؟')) {
+      this.edit = false;
+      this.penaltiesFacade.deletePenalties(Id);
+      this.registerForm.reset();
+    }
   }
   onReset(): void {
     this.edit = false;
@@ -46,21 +47,22 @@ export class PenaltiesComponent implements OnInit {
   }
   onAdd(): void {
     if (this.registerForm.valid) {
-      this.registerForm.value.penaltyTypeName = optionsPenaltyType.find(option => option.value == this.registerForm.value.penaltyTypeId)?.label;
+      this.registerForm.value.penaltyTypeName = optionsPenaltyType.find(
+        (option) => option.value == this.registerForm.value.penaltyTypeId
+      )?.label;
 
-      if(this.edit) {
+      if (this.edit) {
         this.penaltiesFacade.UpdatePenalties(this.registerForm?.value);
         this.onReset();
-      }else{
+      } else {
         this.penaltiesFacade.AddPenalties(this.registerForm?.value);
         this.onReset();
       }
-    }else {
-      if(this.registerForm.value.penaltyTypeId  == 0 || this.registerForm.controls.penaltyTypeId.invalid ){
+    } else {
+      if (this.registerForm.value.penaltyTypeId == 0 || this.registerForm.controls.penaltyTypeId.invalid) {
         this.sharedFacade.showMessage(MessageType.warning, 'عفواً، الرجاء اختر نوع الجزاء  ', ['']);
         return;
-      }
-      else if( this.registerForm.controls.discount.invalid ){
+      } else if (this.registerForm.controls.discount.invalid) {
         this.sharedFacade.showMessage(MessageType.warning, 'عفواً، الرجاء ادخال الخصم', ['']);
         return;
       }
@@ -68,7 +70,9 @@ export class PenaltiesComponent implements OnInit {
   }
   onEdit(penalties: any): void {
     this.registerForm.patchValue(penalties);
-    this.registerForm.value.penaltyTypeName = optionsPenaltyType.find(option => option.value == this.registerForm.value.penaltyTypeId)?.label;
+    this.registerForm.value.penaltyTypeName = optionsPenaltyType.find(
+      (option) => option.value == this.registerForm.value.penaltyTypeId
+    )?.label;
     this.edit = true;
   }
   activate(item): void {
